refactor(tpd): split redis_store.go by topic
redis_store.go was 1824 lines with four largely-independent topic groups
(transport CRUD, bandwidth tracking, uptime/heartbeats, metrics) that
have all evolved separately under the recent perf work. Split along
those lines:

  redis_store.go     (122)  struct, constructor, Close
  redis_transport.go (524)  Register/Deregister, Get*, scanAllTransports,
                              indexes, key helpers, dataToEntry
  redis_bandwidth.go (286)  bandwidth keys, updateBandwidth,
                              GetTransportBandwidth, GetVisorBandwidth,
                              BackupAndCleanOldBandwidth
  redis_uptime.go    (347)  visor + transport heartbeats, daily summaries,
                              timeline timeslot helpers
  redis_metrics.go   (574)  GetNetworkMetrics, GetVisorAggregateMetrics,
                              GetAllTransportMetrics, GetTransportMetricsByIDs,
                              GetTransportMetricsByVisors, buildTransportMetrics,
                              getP2PTransportCounts, getBandwidthFromHash

No behavior change — pure file split. Each new feature should now have
a clear home and PR diffs no longer span unrelated topics.
